linter and separation of library code and tests
